Added QualityType bonuses to DropTables.
QualityType Bonuses allows the DropTables to modify the chances of a particular QualityType being selected for from a given DropTable entry.

Fixed an issue where the QualityType selection code was incorrectly calculating the final selection due to the RAND function overflowing back to 0 inappropriately when the PickChance for a QualityType went negative.