Implementation of Stackable Items.
Setup the ability for ItemDefinitions to describe how an Item can be stacked.

This is managed through a new class type `UItemStackSettings`

Added in functionality for handling Stackables on the ItemInventoryComponent.

Introduced Gold and Potions as Stackables in the Sample Project, with modifications to the Character Inventory for showing your current Gold and providing a mechanism to drop it back on the ground as an ItemDrop.